backend/requirements.txt
fastapi
uvicorn
sqlmodel
pyjwt
bcrypt
python-dotenv
python-multipart
python-jose